test: name test files after the files they cover
Six test files had no source file of their own, so the xx.go/xx_test.go pairing
did not hold: the godoc auditor's tests were spread across audit_edge, render
and repo, its fixer's across docgen and process, and the llms generator's across
gen. Each is now merged into the _test.go matching the file it exercises.

Bodies are unchanged and no test was dropped (104 test/example functions before
and after). The TestWidget and ExampleWidget declarations that appear twice are
fixture source inside string literals, not package-level declarations, so the
merge needed no renaming.
